Safari-611.1.10/LayoutTests/fast/events/ios/rotation/orientationchange-event-listener-on.body-expected.txt

Tests the behavior of the resize / orientationchange event listeners on both the body and the window.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Before rotation
After rotation
PASS eventListenerCalls is 4
PASS successfullyParsed is true

TEST COMPLETE
Rotation test.
